refactor: move footer categories logic inside separate component
